cli: move clean action to separate file
